python/examples/gdp_comparator/requirements.txt
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Cors==5.0.0
flask-talisman==1.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pip-tools==7.4.1
pluggy==1.5.0
pycparser==2.22
pyOpenSSL==24.3.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
tradingeconomics==4.4.1
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
zstandard==0.23.0
